feat(hermes): Move the hermes folder to list of integrations and their specific codebase
