+++ app/integrations/hermes/__init__.py
@@ -9,27 +9,27 @@ entrypoint.
 
 from __future__ import annotations
 
-from app.hermes.agent import DEFAULT_LOG_PATH, HermesAgent, IncidentSink
-from app.hermes.classifier import (
+from app.integrations.hermes.agent import DEFAULT_LOG_PATH, HermesAgent, IncidentSink
+from app.integrations.hermes.classifier import (
     DEFAULT_TRACEBACK_FOLLOWUP_S,
     DEFAULT_WARNING_BURST_THRESHOLD,
     DEFAULT_WARNING_BURST_WINDOW_S,
     IncidentClassifier,
     classify_all,
 )
-from app.hermes.incident import HermesIncident, IncidentSeverity, LogLevel, LogRecord
-from app.hermes.investigation import (
+from app.integrations.hermes.incident import HermesIncident, IncidentSeverity, LogLevel, LogRecord
+from app.integrations.hermes.investigation import (
     build_alert_from_incident,
     run_incident_investigation,
 )
-from app.hermes.parser import parse_log_line
-from app.hermes.sinks import (
+from app.integrations.hermes.parser import parse_log_line
+from app.integrations.hermes.sinks import (
     InvestigationBridge,
     TelegramSink,
     TelegramSinkConfig,
     make_telegram_sink,
 )
-from app.hermes.tailer import FileTailer
+from app.integrations.hermes.tailer import FileTailer
 
 __all__ = [
     "DEFAULT_LOG_PATH",
