+++ app/integrations/hermes/sinks.py
@@ -52,9 +52,9 @@ from concurrent.futures import TimeoutError as FutureTimeoutError
 from dataclasses import dataclass
 from typing import Final
 
-from app.hermes.agent import IncidentSink
-from app.hermes.errors import InvestigationOutcome
-from app.hermes.incident import HermesIncident, IncidentSeverity, LogRecord
+from app.integrations.hermes.agent import IncidentSink
+from app.integrations.hermes.errors import InvestigationOutcome
+from app.integrations.hermes.incident import HermesIncident, IncidentSeverity, LogRecord
 from app.watch_dog.alarms import AlarmDispatcher
 
 logger = logging.getLogger(__name__)
