+++ docs/hermes.mdx
@@ -149,6 +149,6 @@ Contributors add scenarios under `tests/synthetic/hermes/` (fixtures, `scenario.
 
 ## Further reading
 
-- Implementation lives under `app/hermes/` in the OpenSRE repository (tailer, parser, classifier, correlator, sinks, CLI wiring).
+- Implementation lives under `app/integrations/hermes/` in the OpenSRE repository (tailer, parser, classifier, correlator, sinks, CLI wiring).
 - If you also run [OpenClaw](/openclaw) alongside Hermes, that page covers the OpenClaw bridge (context lookup and RCA write-back).
 - For false-positive risks when rules match both `message` and raw line text, see [issue #1874](https://github.com/Tracer-Cloud/opensre/issues/1874).
