+++ tests/synthetic/hermes/__init__.py
@@ -2,7 +2,7 @@
 
 Each scenario lives in a numbered subdirectory and bundles a real-world
 ``errors.log`` fixture plus a YAML answer key describing the incidents
-the :mod:`app.hermes` classifier is expected to emit. See
+the :mod:`app.integrations.hermes` classifier is expected to emit. See
 ``tests/synthetic/hermes/README.md`` for the schema and how to add a
 new case.
 """
