+++ tests/tools/test_hermes_logs_tool.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import pytest
 
-from app.hermes.poller import HermesLogCursor
+from app.integrations.hermes.poller import HermesLogCursor
 from app.tools.HermesLogsTool import get_hermes_logs
 
 _LINES = [
